+++ includes/class.llms.ajax.php
@@ -46,7 +46,7 @@ class LLMS_AJAX {
 		$response = LLMS_AJAX_Handler::$request['action']( $request );
 
 		if ( $response instanceof WP_Error ) {
-			$this->send_error( $response );
+			self::send_error( $response );
 		}
 
 		wp_send_json_success( $response );
@@ -131,7 +131,7 @@ class LLMS_AJAX {
 	 *
 	 * @param WP_Error $error
 	 */
-	private function send_error( $error ) {
+	private static function send_error( $error ) {
 		wp_send_json(array(
 			'code' => $error->get_error_code(),
 			'message' => $error->get_error_message(),
@@ -145,6 +145,7 @@ class LLMS_AJAX {
 
 		$ajax_events = array(
 			'get_courses'				=> false,
+			'get_memberships'			=> false,
 			'get_course_tracks'			=> false,
 			'get_sections'				=> false,
 			'get_lesson'				=> false,
@@ -315,6 +316,28 @@ class LLMS_AJAX {
 			'post_status'   => 'publish',
 
 		 );
+
+		$postslist = get_posts( $args );
+
+		echo json_encode( $postslist );
+
+		die();
+	}
+
+	/**
+	 * Return array of memberships (id => name)
+	 *
+	 * @param string
+	 * @return array
+	 */
+	public function get_memberships() {
+
+		$args = array(
+				'post_type' 	=> 'llms_membership',
+				'nopaging' 		=> true,
+				'post_status'   => 'publish',
+
+		);
 
 		$postslist = get_posts( $args );