+++ CHANGELOG.md
@@ -1,13 +1,32 @@
 == Changelog ==
 
-= v2.2.4 - 2016/03/15 =
------------------------
+= v2.3.0-1 - 2016/03/22 =
+-------------------------
 
+##### Engagements Refactoring (lots of bugfixes, performance improvements, more hook & filter friendly)
+
++ We've completely rewritten the LifterLMS Engagement Handler methods (`class LLMS_Engagements`) and added some new engagement actions.
++ The rewrite unifies engagement handling into one function that can be easily hooked into by plugin and theme developers.
++ We've moved any engagment related data out of the main `LifterLMS` class
++ Fixed the broken engagement delay functionality which now runs of `wp_schedule_single_event`. This makes the function more reliable and also keeps it within the traditional WordPress architecture.
 + Added an additional check before sending emails or triggering any engagments that will prevent the achievement from being awarded or the email from being sent if the post is in not published. This fixes an issue that caused emails in the trash from still being emailed.
++ Removed the unused `LLMS_Engagements` class and file
++ Added two new engagment trigger events "Membership Purchased" and "Course Purchased"
++ Deprecated actions -- Removes some redundancy because the triggering actions (`lifterlms_course_completed` triggered the notification action, instead `lifterlms_course_completed` simply triggers the engagement now).
+  + `lifterlms_lesson_completed_notification`
+  + `lifterlms_section_completed_notification`
+  + `lifterlms_course_completed_notification`
+  + `lifterlms_course_track_completed_notification`
+  + `lifterlms_course_completed_notification`
+  + `lifterlms_user_purchased_product_notification`
+  + `lifterlms_created_person_notification`
+
+##### Bug and Issue fixes
+
++ Fixed a function that should have been called statically in `LLMS_Ajax` class
 + Fixed a ton of issues related to the triggering of engagements and cleaned up a lot of classes and functions associated with them.
 + Properly instatiate `LifterLMS` singleton via LLMS() function and prevent direct instantiation of the class via `new LifterLMS()`.
 + Removed the deprecated 'class.llms.email.person.new.php' file as it was rendered useless a long time ago and caused some duplicate emails.
-+ Removed the unused `LLMS_Engagements` class and file
 
 
 = v2.2.3 - 2016/03/15 =